Fix(RefreshToken): Create a new variable for `parsed_request_args`
Prior to this commit, we were saving a merge of user arguments and default arguments in a variable that held user arguments. This led to an issue where, after refreshing the token and making another API call, the old cookie saved in the user arguments variable was inadvertently used, so the user had to refresh the page to call API with the new cookie.
By storing the merged arguments in a new variable, we ensure that `request_args` always contains user-defined arguments.